+++ brownie/exceptions.py
@@ -61,22 +61,17 @@ class VirtualMachineError(Exception):
         revert_msg: The returned error string, if any.
         source: The contract source code where the revert occured, if available."""
 
-    revert_msg = ""
-    source = ""
-
     def __init__(self, exc: Any) -> None:
         if type(exc) is not dict:
             try:
                 exc = eval(str(exc))
             except SyntaxError:
                 exc = {"message": str(exc)}
-        if len(exc["message"].split("revert ", maxsplit=1)) > 1:
-            self.revert_msg = exc["message"].split("revert ")[-1]
-        if "source" in exc:
-            self.source = exc["source"]
-            super().__init__(exc["message"] + "\n" + exc["source"])
-        else:
-            super().__init__(exc["message"])
+        self.revert_msg = msg = exc["message"]
+        self.source = exc.get("source", "")
+        if self.source:
+            msg = f"{msg}\n{self.source}"
+        super().__init__(msg)
 
 
 class EventLookupError(LookupError):