+++ brownie/project/build.py
@@ -53,19 +53,27 @@ class Build:
             for k, v in pcMap.items()
             if v["op"] in {"REVERT", "INVALID"} or "jump_revert" in v
         ):
-            if "fn" not in data or "first_revert" in data:
-                _revert_map[pc] = False
-                continue
-            data["dev"] = ""
-            try:
-                revert_str = self._sources.get(data["path"])[data["offset"][1] :]
-                revert_str = revert_str[: revert_str.index("\n")]
-                revert_str = revert_str[revert_str.index("//") + 2 :].strip()
-                if revert_str.startswith("dev:"):
-                    data["dev"] = revert_str
-            except (KeyError, ValueError):
-                pass
-            revert = (data["path"], tuple(data["offset"]), data["fn"], data["dev"], self._sources)
+            if "dev" not in data:
+                if "fn" not in data or "first_revert" in data:
+                    _revert_map[pc] = False
+                    continue
+                try:
+                    revert_str = self._sources.get(data["path"])[data["offset"][1] :]
+                    revert_str = revert_str[: revert_str.index("\n")]
+                    revert_str = revert_str[revert_str.index("//") + 2 :].strip()
+                    if revert_str.startswith("dev:"):
+                        data["dev"] = revert_str
+                except (KeyError, ValueError):
+                    pass
+
+            msg = "" if data["op"] == "REVERT" else "invalid opcode"
+            revert = (
+                data["path"],
+                tuple(data["offset"]),
+                data["fn"],
+                data.get("dev", msg),
+                self._sources,
+            )
 
             # do not compare the final tuple item in case the same project was loaded twice
             if pc not in _revert_map or (_revert_map[pc] and revert[:-1] == _revert_map[pc][:-1]):